# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GoalBet Virtual Football Betting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /user/bets:
    get:
      operationId: getUserBets
      tags: [user]
      summary: Get user bet history
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Bet history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetListResponse"
...
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get matches
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, completed]
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /bets:
    post:
      operationId: placeBet
      tags: [bets]
      summary: Place a bet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBetRequest"
      responses:
        "201":
          description: Bet placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/matches:
    post:
      operationId: adminCreateMatch
      tags: [admin]
      summary: Create a match
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
...
  /admin/bets:
    get:
      operationId: adminGetBets
      tags: [admin]
      summary: Get all bets
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: matchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetListResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required: [message]
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        totalBets:
          type: number
        totalWins:
          type: number
        totalWinnings:
          type: number
        createdAt:
          type: string
      required: [id, username, email, balance, role, status]
...
    OddsObject:
      type: object
      properties:
        home:
          type: number
        draw:
          type: number
        away:
          type: number
      required: [home, draw, away]
    MatchEvent:
      type: object
      properties:
        minute:
          type: number
        type:
          type: string
        team:
          type: string
        player:
          type: string
        description:
          type: string
      required: [minute, type, team]
    MatchResponse:
      type: object
      properties:
        id:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeScore:
          type: number
        awayScore:
          type: number
        status:
          type: string
          enum: [upcoming, betting_open, live, completed, cancelled]
        odds:
          $ref: "#/components/schemas/OddsObject"
        minute:
          type: number
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEvent"
        scheduledAt:
          type: string
        bettingClosesAt:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        result:
          type: string
          enum: [home, draw, away]
        totalBets:
          type: number
        totalBetAmount:
          type: number
      required: [id, homeTeam, awayTeam, homeScore, awayScore, status, odds]
    MatchListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/MatchResponse"
        total:
          type: number
      required: [matches, total]
    PlaceBetRequest:
      type: object
      properties:
        matchId:
          type: string
        outcome:
          type: string
          enum: [home, draw, away]
        amount:
          type: number
      required: [matchId, outcome, amount]
    BetResponse:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        userId:
          type: string
        username:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        outcome:
          type: string
          enum: [home, draw, away]
        amount:
          type: number
        odds:
          type: number
        potentialWinnings:
          type: number
        status:
          type: string
          enum: [pending, won, lost, refunded]
        actualWinnings:
          type: number
        createdAt:
          type: string
        settledAt:
          type: string
      required: [id, matchId, userId, outcome, amount, odds, potentialWinnings, status, createdAt]
    BetListResponse:
      type: object
      properties:
        bets:
          type: array
          items:
            $ref: "#/components/schemas/BetResponse"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required: [bets, total]
...
    CreateMatchRequest:
      type: object
      properties:
        homeTeam:
          type: string
        awayTeam:
          type: string
        odds:
          $ref: "#/components/schemas/OddsObject"
        scheduledAt:
          type: string
        bettingWindowMinutes:
          type: number
      required: [homeTeam, awayTeam, odds]
...